requirements.txt
pyautogui==0.9.54
pyobjc-framework-Quartz==10.3.1
